+++ pyogrio/geopandas.py
@@ -190,6 +190,7 @@ def read_dataframe(
     import pandas as pd
     import geopandas as gp
     from geopandas.array import from_wkb
+    import shapely  # if geopandas is present, shapely is expected to be present
 
     path_or_buffer = _stringify_path(path_or_buffer)
 
@@ -197,13 +198,14 @@ def read_dataframe(
         use_arrow = bool(int(os.environ.get("PYOGRIO_USE_ARROW", "0")))
 
     read_func = read_arrow if use_arrow else read
+    gdal_force_2d = False if use_arrow else force_2d
     result = read_func(
         path_or_buffer,
         layer=layer,
         encoding=encoding,
         columns=columns,
         read_geometry=read_geometry,
-        force_2d=force_2d,
+        force_2d=gdal_force_2d,
         skip_features=skip_features,
         max_features=max_features,
         where=where,
@@ -237,6 +239,8 @@ def read_dataframe(
             return pd.DataFrame()
         elif geometry_name in df.columns:
             df["geometry"] = from_wkb(df.pop(geometry_name), crs=meta["crs"])
+            if force_2d:
+                df["geometry"] = shapely.force_2d(df["geometry"])
             return gp.GeoDataFrame(df, geometry="geometry")
         else:
             return df