+++ CHANGES.md
@@ -24,6 +24,7 @@
     some drivers; this has been fixed in GDAL 3.8.0.
 -   Removed warning when no features are read from the data source (#299).
 -   Add support for `fids` filter with `use_arrow=True` (#304)
+-   Add support for `force_2d=True` with `use_arrow=True` in `read_dataframe` (#300).
 
 ### Other changes